==> P2_tidy_data.Rmd <==
---
title: "Final Project"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
## Visual Attention and Learning
### Loading and tidying the data

### Charlie Heller and Jyoti Tyagi
### Math 630

First, load the libraries we'll need for this task.
```{r echo=TRUE, eval=FALSE}
library(readxl)
library(tidyverse)
```

```{r echo=TRUE, eval=TRUE, include=FALSE}
library(readxl)
library(tidyverse)
```

#### Now Let's get started by loading the data
* The data was provided in an excel document with multiple sheets. Let's read this in as a list of dataframes.
* The method we will use requires the package readxl which we have loaded above.
* This is a slightly complicated task. We'll right a function to automate it.

```{r echo=TRUE, eval=TRUE}
read_excel_allsheets <- function(filename, tibble = TRUE) {
    sheets <- readxl::excel_sheets(filename)
    x <- lapply(sheets, function(X) readxl::read_excel(filename, sheet = X))
    if(!tibble) x <- lapply(x, as.data.frame)
    names(x) <- sheets
    x
}
```

* Now, let's call the function to load our data

```{r echo=TRUE, eval=TRUE}
data_tables <- read_excel_allsheets(here::here("Data", "vis_att_data.xlsx"))
```
Cool, as we can see below, we now have a data table for each of the individual sheets in the excel file.

Before we get to tidying, let's check out the first sheet we loaded:
```{r echo=TRUE, eval=TRUE}
glimpse(data_tables[1]$`Annotation for all tabs`)
```
Lucky us! They supplied a codebook with the data. All we need to do is add a columns that specifies the variable type. This is also convenient because we happend to know all the data type are continuous, numerical variables.

```{r echo=TRUE, eval=TRUE}
codebook <- data_tables[1]$`Annotation for all tabs` %>%
  mutate(data_type = "numerical continuous")
glimpse(codebook)
```
Cool, now we've got a codebook that tells us everything about our data. We'll go ahead and save that to a csv here:
```{r echo=TRUE, eval=TRUE}
write_csv(codebook, here::here("Data/codebook.csv"))
```

Okay, now for tidying data there's not a whole lot to be done here.From a visual inspection of the data, we determined that some values were missing. These values were filled in as "absent" by the researchers. For the sake of R, let's go trough each of our data frames, search for absent entries and replace with NULL.

```{r echo=TRUE, eval=TRUE}
# loop over all data_tables (excpet the first), ID absent entries and mutate these to NULL
# create copy of data
tidy_data <- rlang::duplicate(data_tables)
for (df in 2:length(data_tables)) {
    name <- names(data_tables[df])
    df_ = get(name, data_tables[df])
    df_[df_=="absent"] <- NA
    df_[df_=="Absent"] <- NA
    tidy_data[[name]] = df_
}
tidy_data$`Summary Data File`$`Sparse-TimeSelfDistr`
data_tables$`Summary Data File`$`Sparse-TimeSelfDistr`
```
Nice, we see that we've replaced all values of "absent" or "Absent" with NA. This is really all our data needed to tidy. We'll save it into our data folder for later use.

```{r echo=TRUE, eval=TRUE}
library(xlsx)
for (sheet in 1:length(tidy_data)) {
  name <- names(tidy_data[sheet])
  if (sheet==1){
  write.xlsx(get(name, tidy_data), here::here("Data/tidy_vis_data.xlsx"), sheetName=name)
  } else {
    wb <- loadWorkbook(here::here("Data/tidy_vis_data.xlsx"))
    sheet <- createSheet(wb, sheetName = name)

    addDataFrame(get(name, tidy_data), sheet, startRow=1, startColumn=1,
                 colnames=names(names(get(name, tidy_data))))
    saveWorkbook(wb, here::here("Data/tidy_vis_data.xlsx"))
  }
}
```
